Refactor global declarations.
I discovered in #8 that you can declare a variable as global at the top rather than in each method. This would be useful for cleaning up the ones that use xScale, yScale in multiple places.

All colors are now read into global variables so that if users have trouble, we can quickly check which pixels aren't matching and why.

Resolves #10.